Refresh
My wish was to make a function that reads the data, and whenever the data is updated, a new function will remove all data from the map and add the new one. This way the map will automatically stay updated, and the user will not have to refresh the page.
None of the things I've tried has worked so far, so I'm currently using a jalla refresh-function that refreshes the page automatically every 10 seconds, keep the visualized data freshhh . Not optimal.